feature/productOption/variantId-error-#22_0 Syncing our custom components closer to RCs
