+++ imports/plugins/custom/product-detail-option/client/components/index.js
@@ -4,6 +4,6 @@ export { default as VariantList } from "./variantList";
 export { default as ProductMetadata } from "./metadata";
 export { default as ProductField } from "./productField";
 export { default as ChildVariant } from "./childVariant";
-export { default as ProductDetail } from "./productDetail";
+export { default as ProductOptionComponent } from "./productOptionComponent";
 export { default as AddToCartButton } from "./addToCartButton";
 export { default as ProductNotFound } from "./productNotFound";
