+++ imports/plugins/custom/product-detail-option/client/containers/index.js
@@ -1,4 +1,4 @@
 export { default as SocialContainer } from "./socialContainer";
 export { default as PublishContainer } from "./publishContainer";
 export { default as VariantListContainer } from "./variantListContainer";
-export { default as ProductOptionContainer } from "./productDetailContainer";
+export { default as ProductOptionContainer } from "./productOptionContainer";
